requirements.txt
# KubeGuard FastMCP Server Requirements

# FastMCP Framework
fastmcp>=0.9.0

# Data handling
pydantic>=2.0.0
pyyaml>=6.0

# LLM Client support
openai>=1.0.0
anthropic>=0.3.0
httpx>=0.24.0

# Kubernetes manifest handling
kubernetes>=28.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0